web - restructure components/common dir
