+++ web/src/components/common/constants.ts
@@ -1,0 +1,5 @@
+// Disable this list for now to show UEFA Euros 2024 games first
+// export const TOP_COMPS_IDS: number[] = [2, 3, 39, 45, 48, 61, 78, 135, 140, 253, 848];
+
+// UEFA Euros 2024 & Copa America
+export const TOP_COMPS_IDS: number[] = [4, 9];
