+++ web/src/components/common/types.ts
@@ -1,9 +1,3 @@
 export interface IProps {
     children?: React.ReactNode;
 }
-
-// Disable this list for now to show UEFA Euros 2024 games first
-// export const TOP_COMPS_IDS: number[] = [2, 3, 39, 45, 48, 61, 78, 135, 140, 253, 848];
-
-// UEFA Euros 2024 & Copa America
-export const TOP_COMPS_IDS: number[] = [4, 9];
