+++ web/src/components/features/match/match-events/match-events-item/match-events-item.tsx
@@ -1,5 +1,5 @@
 import { ApiFootballLogoComponent } from "@/components/common/api-football-logo/api-football-logo";
-import { IMatchEvent, MatchEventType } from "@/components/common/api-football-response";
+import { IMatchEvent, MatchEventType } from "@/components/common/responses/api-football";
 import { MatchToolbox } from "@/components/common/toolbox/match";
 import { Card, CardContent } from "@/components/ui/card";
 import { NavLink } from "react-router-dom";
