+++ web/src/components/features/match/match-lineups/match-lineups.tsx
@@ -1,4 +1,4 @@
-import { IMatchLineup } from "@/components/common/api-football-response";
+import { IMatchLineup } from "@/components/common/responses/api-football";
 import { MatchLineupsFormationsComponent } from "./match-lineups-formations/match-lineups-formations";
 import { MatchLineupsSquadsComponent } from "./match-lineups-squads/match-lineups-squads";
